init base class. ensemble selection train file fix on cawpe paper branch copied over to this one. tests confirmed reproducing and ready to run locally
