+++ src/main/java/weka_uea/classifiers/ensembles/AbstractEnsemble.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright (C) 2019 xmw13bzu
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package weka_uea.classifiers.ensembles;
+
+import weka.classifiers.AbstractClassifier;
+import weka.core.Instances;
+
+/**
+ *
+ * @author James Large (james.large@uea.ac.uk)
+ */
+public class AbstractEnsemble extends AbstractClassifier {
+
+    @Override
+    public void buildClassifier(Instances data) throws Exception {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
+}
